backend/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
python-dotenv
google-generativeai
